Refactor GetUserAsync to use email parameter in URL
Updated the route for the GetUserAsync method to include an email parameter in the URL path. Changed the method signature to accept a string email parameter instead of a JObject form. Commented out the code that used jsonObject to get the email and now directly use the email parameter to retrieve user information. These changes simplify the method by directly passing the email as a parameter in the URL.