Comment out FlightCrewId assignment in ProcessFile.cs
The `FlightCrewId` property assignment in the `ProcessFile.cs` file within the `CrewMobileApi.Business` namespace has been commented out. This change means that the `FlightCrewId` will no longer be set when creating the object, instead the value will be assigned by the SQL Server identity configured for that column in the database.